move optimization to tweek-client
