+++ js/tweek-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tweek-client",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "description": "Tweek client for JavaScript",
   "author": "Soluto",
   "license": "MIT",
