+++ js/tweek-client/src/index.ts
@@ -4,3 +4,4 @@ export { default as TweekClientWithFallback } from './TweekClientWithFallback';
 export { default as createTweekClient } from './createTweekClient';
 export { default as createTweekClientWithFallback } from './createTweekClientWithFallback';
 export { default as VersionWatcher } from './VersionWatcher';
+export { isScanKey } from './utils';
