+++ js/tweek-local-cache/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tweek-local-cache",
-  "version": "0.4.1",
+  "version": "0.4.0",
   "description": "Local cache to be used with tweek-client",
   "author": "Soluto",
   "license": "MIT",
