+++ js/tweek-local-cache/src/tweek-repository.ts
@@ -12,7 +12,6 @@ import {
   isNullOrUndefined,
   isScanKey,
   once,
-  optimizeInclude,
   partitionByIndex,
   snakeToCamelCase,
 } from './utils';
@@ -278,7 +277,7 @@ export default class TweekRepository {
       flatten: true,
       casing: TweekCasing.snake,
       context: this._context,
-      include: optimizeInclude(keysToRefresh),
+      include: keysToRefresh,
     };
 
     return this._client
